requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
transformers
accelerate
pymupdf
numpy